requirements.txt
SQLAlchemy~=1.4.45
alembic~=1.9.0
fastapi~=0.88.0
boto3~=1.26.29
Authlib~=1.2.0
pydantic~=1.10.2
passlib~=1.7.4
starlette~=0.22.0
loguru~=0.6.0
factory-boy~=3.2.1
python-jose~=3.3.0
pytest~=7.2.0
mutagen~=1.46.0